requirements.txt
boto3==1.10.8
botocore==1.13.8
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-rest-knox==4.1.0
django-storages==1.7.2
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
pycparser==2.19
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
wincertstore==0.2
